requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopage==0.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
catboost==1.0.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorlog==6.6.0
cycler==0.11.0
datasets==2.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
filelock==3.6.0
flake8==4.0.1
flatbuffers==2.0
fonttools==4.31.1
frozenlist==1.3.0
fsspec==2022.2.0
gast==0.5.3
gensim==4.1.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
lazy-object-proxy==1.7.1
libclang==13.0.0
lightgbm==3.3.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.19.0
multidict==6.0.2
multiprocess==0.70.12.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.4
notebook==6.4.10
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pdoc3==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.6.0
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
singledispatch==3.7.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
SQLAlchemy==1.4.32
stevedore==3.5.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.13.3
xgboost==1.5.2
xxhash==3.0.0
yarl==1.7.2
zipp==3.7.0
